extract methods from Tetris to CommandLineAnalyst
